requirements-crewai.txt
# Core CrewAI with specific versions that work together
crewai==0.28.8
crewai-tools==0.1.6

# Google Gemini (instead of OpenAI)
google-generativeai==0.3.0

# LangChain components (pinned to compatible versions)
langchain==0.1.12
langchain-community==0.0.28
langchain-google-genai==0.0.11

# Utilities
python-dotenv==1.0.0
pydantic==2.6.1
typing-extensions==4.8.0

# Web scraping for research
beautifulsoup4==4.12.3
requests==2.31.0

# For API (optional)
fastapi==0.110.0
uvicorn==0.27.1
